face_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings
boto3
deepface
opencv-python
tensorflow
numpy
scipy
Pillow